Rename tx_hash_hex to hex_tx_hash
